refactoring code footer
