+++ src/layout/sections/projects/project/Project.tsx
@@ -8,7 +8,7 @@ export type ProjectPropsType = {
     src: string
 }
 
-export const Project = (props: ProjectPropsType) => {
+export const Project: React.FC<ProjectPropsType> = (props: ProjectPropsType) => {
     return (
         <S.Project>
             <S.Image src={props.src} alt=""/>
