+++ src/layout/sections/skills/skill/Skill.tsx
@@ -10,7 +10,7 @@ export type SkillPropsType = {
     viewBox?: string
 }
 
-export const Skill = (props: SkillPropsType) => {
+export const Skill: React.FC<SkillPropsType> = (props: SkillPropsType) => {
     return (
         <S.Skill>
             <S.IconWrapper>
